Update SearchProviders page filter column
SearchProviders page now no longer uses mock data. Added a fetch to the
ServiceQuestionService to grab the questions for a given service id. The
searchproviders page populates the filters from the questions retrieved
from there, after doing some basic parsing on the choices in the
questions, transforming from a comma separated string to an array.